Fix Mole bad active variable reset
Fixed a bug when poping a Mole, the Mole "active" bool wouldn't reset
correctly, leading to a crash of the game after a certain time due to an
infinite while loop in the WallManager. Now the "active" bool has been
removed since the "state" enum fills the same job, fixing the bug.