Add Pointer smooth and auto-aim
Added a smoothing option to the Pointer, which stabilises the laser in
case of shaky movements. Two aiming assistance were also added, a
snapping assistance, which snaps the laser to the center of the closest
Mole, and a magnetizing assistance, which pulls the laser towards the
closest Mole. Also added corresponding logs.