Add profile UI and profile managment
Added profiles and their corresponding UI. A profile is used to store
certain parameters/properties to facilitate test presets and follow-up.
A corresponding UI has been added, allowing to select, delete and create
a profile. It is mandatory to select a profile to start the game.
Profiles are locally saved in the
AppData/LocalLow/[project]/UserProfiles folder as *.uprofile files,
using the *.ini or *.cgf format style.