Initial Support for Physical Mirror Mode
This implementation is the first attempt at a
physical mirror mode. The implementation is simple
but has some limitations regarding the perception
quality.

In the current state, to use the physical mirror mode
properly, one would need to use Whack-A-Mole patterns
which only create moles on the right-side of the wall.
The Physical Mirror currently dont support switching
side.